Implement Kafka communication following Saga Choreography design pattern. Added handling flow for creating properties with invalid post services.

+++ backend/services/property-service/src/main/java/com/koi151/msproperty/controller/RoomController.java
@@ -18,7 +18,7 @@ public class RoomController {
 
     @PostMapping(value = "/{propertyId}/rooms")
     public ResponseEntity<ResponseData> createRoom(
-            @PathVariable("propertyId") Integer propertyId,
+            @PathVariable("propertyId") Long propertyId,
             @RequestBody RoomCreateUpdateRequest request
     ) {
