+++ backend/services/property-service/src/main/java/com/koi151/msproperty/model/dto/RoomDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 public class RoomDTO {
     private long roomId;
-    private int propertyId;
+    private long propertyId;
     private String roomType;
     private int quantity;
 }
